Harden the Mach ABI contract and mirror the module split in the tests
Review-battery fixes on top of the macOS available-RAM fix:

- Pin HOST_VM_INFO64_REV1_COUNT=38 with a sizeof self-consistency guard.
  The kernel refuses only counts below the lowest legal revision boundary;
  an oversized or off-boundary count is silently clamped to a boundary it
  knows and still returns KERN_SUCCESS (verified empirically: count=100 is
  accepted and clamped), so a drifted struct cannot be detected at call
  time and must be caught statically.
- Guard mach_port_deallocate in the finally block: a leaked send right is
  preferable to discarding a good reading or masking the body's real error.
- Pin the fake libSystem's ABI expectations (flavour 4, 38-word count),
  hard-coded independently of the module constants, so drift is visible on
  Linux CI where only the fake ever answers.
- Bound the unmocked darwin test above by total installed RAM - the one
  assertion distinguishing available from total against the real kernel.
- Assert the failed probe queries the kernel exactly once, and add the
  optimistic-bound caveat (inactive includes compression-reclaimable dirty
  pages; compressor memory is not subtracted) to docstring and spec.
- Move all host-observation tests to a new tests/test_host_memory.py,
  mirroring the module split; merge the duplicated negative-cgroup tests;
  re-key the two relocated skipif fingerprints and regenerate the ledgers.
- Documentation accuracy: scope the recorded-reason claim to the RAM
  probes (available_vram_bytes reports total VRAM, reasonless None), add
  _host_memory.py to the high-level capability list, and de-Linux-ify its
  available-memory bullet.
