+++ specs/execution-engine/low-level.md
@@ -21,7 +21,7 @@
 | `src/haute/_graph_utils.py` | Pure-function graph helpers decoupled from the Pydantic models: `build_parents_of`, `upstream_node_ids`, `_sanitize_func_name`, `edge_input_name` (the single edge→input-name derivation: apiInput-frame edge → its frame label verbatim, else sanitised source-node label; consumed by the executor, codegen, projection, and the deploy scorer so all four agree byte-for-byte), `build_instance_mapping`, `resolve_orig_source_names`, and edge-id construction. |
 | `src/haute/_worker_isolation.py` | `run_isolated_worker()` — spawn a child process for one function call with an optional address-space resource cap, timeout, and cooperative stop-reason polling; typed error hierarchy for every terminal state. |
 | `src/haute/chunking.py` | `ChunkPlanRequest`/`chunk_plan()` (proves a graph suffix is chunk-safe, sizes chunks from projected target width, and rejects an over-budget single target row), `iter_chunked_frames()`/`run_chunked_reduce()`/`collect_chunked()` (the serial runner), the per-`NodeType` `ChunkCapability` registry, and the AST-based row-local user-code whitelist. |
-| `src/haute/_host_memory.py` | Host memory observation: `available_ram_bytes()`/`available_vram_bytes()` (per-platform probes behind one shared result contract, including Linux cgroup v2/v1 headroom clamping). Never fabricates capacity — each probe reports a real measurement or a recorded failure reason. Owns the nvidia-smi subprocess chokepoint. |
+| `src/haute/_host_memory.py` | Host memory observation: `available_ram_bytes()` (per-platform probes behind one shared result contract, including Linux cgroup v2/v1 headroom clamping — each probe reports a real measurement or a recorded failure reason, never fabricated capacity) and `available_vram_bytes()` (*total* installed GPU VRAM via nvidia-smi, or nothing when no GPU is detected). Owns the nvidia-smi subprocess chokepoint. |
 | `src/haute/_ram_estimate.py` | Workload-side estimation: `estimate_safe_training_rows()` (parquet-metadata-based peak-memory estimate and downsample decision), `estimate_gpu_vram_bytes()`, and the `MaterialisationEstimate` contract consumed by strategy planning. It imports graph models directly from `_types.py` so admission and route cold imports do not re-enter the execution facade. |
 
 ## Key types and data structures
@@ -125,8 +125,11 @@
   contribute no separate term. `purgeable_count` is deliberately **excluded**:
   purgeable is an attribute of pages that remain on the active/inactive queues
   rather than a disjoint pool, so adding it would double-count pages already
-  inside `inactive_count` and over-admit work. Total installed RAM is never
-  substituted for availability.
+  inside `inactive_count` and over-admit work. The value is an optimistic
+  bound — `inactive_count` includes dirty pages reclaimable only through
+  compression or swap, and compressor-held memory is not subtracted; the
+  admission OS reserve and safety factor absorb that gap. Total installed RAM
+  is never substituted for availability.
 - **Unobservable availability** — a probe that fails records its reason, and
   the single `available_ram_unavailable` warning reports every attempted
   source's reason (a source is attempted exactly when it reports a reason), so
@@ -652,7 +655,10 @@ Tests live in `tests/` (flat layout, no package-per-component subdirectories).
 - **`test_graph_utils.py`** — `_execute_lazy` re-export surface,
   `_sanitize_func_name`, `ancestors`/`topo_sort_ids`
   via the `graph_utils` facade.
-- **`test_ram_estimate.py`** — RAM/VRAM probing across platforms (mocked),
+- **`test_host_memory.py`** — RAM/VRAM probing across platforms: mocked
+  probes for every source and failure path (driving the real ctypes
+  structures), plus one darwin-gated unmocked real-kernel assertion.
+- **`test_ram_estimate.py`** —
   source-metadata resolution (including edge-join key coalescing), and the
   downsample decision. Per-port JSON `apiInput` sizing is exercised against a
   cache built by the same writer the engine reads — each emitted table sized
